Add /users/me endpoint with JWT
Now the API can authenticate with the /token endpoint, which receives
email and password, and returns the JWT token.

The /users/me endpoint requires the token in the header in order to
return the currently logged-in user data.

The next step is to complete the following endpoints:
POST /logout
PUT /users/me